# api/rebump.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/rebump.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/bump": {
      "post": {
        "operationId": "CreateBump",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiBump"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateBumpRequest"
            }
          }
        ],
        "tags": [
          "Rebump"
        ]
      }
    },
    "/bump:nearby": {
      "get": {
        "operationId": "GetNearby",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiBump"
            }
          }
        },
        "parameters": [
          {
            "name": "location.latitude",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "location.longitude",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          }
        ],
        "tags": [
          "Rebump"
        ]
      }
    }
  },
  "definitions": {
    "BumpSecret": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        }
      }
    },
    "apiBump": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "location": {
          "$ref": "#/definitions/apiLocation"
        },
        "url": {
          "type": "string"
        },
        "secret": {
          "$ref": "#/definitions/BumpSecret"
        }
      }
    },
    "apiCreateBumpRequest": {
      "type": "object",
      "properties": {
        "location": {
          "$ref": "#/definitions/apiLocation"
        }
      }
    },
    "apiLocation": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "double"
        },
        "longitude": {
          "type": "number",
          "format": "double"
        }
      }
    }
  }
}